third_party/blink/web_tests/fast/forms/select/customizable-select/disallowed-select-descendants-console-message-expected.txt

CONSOLE ERROR: An element which is not allowed in the content model of the <select> element was found within a <select> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology. If using disallowed elements for layout structure and styling, consider using the allowed <div> element instead. Any text existing within the <select> element should either be removed or relocated to a valid element that allows text descendants, e.g., an <optgroup> with a <legend> element or <option> elements.
CONSOLE ERROR: An element which is not allowed in the content model of the <select> element was found within a <select> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology. If using disallowed elements for layout structure and styling, consider using the allowed <div> element instead. Any text existing within the <select> element should either be removed or relocated to a valid element that allows text descendants, e.g., an <optgroup> with a <legend> element or <option> elements.
CONSOLE ERROR: Non-phrasing content was found within an <option> element. The <option> element allows only non-interactive phrasing content, text, and <div> elements as its children. The semantics of non-phrasing content elements do not make sense as children of an <option>, and such semantics will largely be ignored by assistive technology since they are inappropriate in this context. Consider removing or changing such elements to one of the allowed phrasing content elements.
CONSOLE ERROR: Non-phrasing content was found within an <option> element. The <option> element allows only non-interactive phrasing content, text, and <div> elements as its children. The semantics of non-phrasing content elements do not make sense as children of an <option>, and such semantics will largely be ignored by assistive technology since they are inappropriate in this context. Consider removing or changing such elements to one of the allowed phrasing content elements.
CONSOLE ERROR: Non-phrasing content was found within an <option> element. The <option> element allows only non-interactive phrasing content, text, and <div> elements as its children. The semantics of non-phrasing content elements do not make sense as children of an <option>, and such semantics will largely be ignored by assistive technology since they are inappropriate in this context. Consider removing or changing such elements to one of the allowed phrasing content elements.
CONSOLE ERROR: An element which is not allowed in the content model of the <select> element was found within a <select> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology. If using disallowed elements for layout structure and styling, consider using the allowed <div> element instead. Any text existing within the <select> element should either be removed or relocated to a valid element that allows text descendants, e.g., an <optgroup> with a <legend> element or <option> elements.
CONSOLE ERROR: An element which is not allowed in the content model of the <select> element was found within a <select> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology. If using disallowed elements for layout structure and styling, consider using the allowed <div> element instead. Any text existing within the <select> element should either be removed or relocated to a valid element that allows text descendants, e.g., an <optgroup> with a <legend> element or <option> elements.
CONSOLE ERROR: An element which is not allowed in the content model of the <select> element was found within a <select> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology. If using disallowed elements for layout structure and styling, consider using the allowed <div> element instead. Any text existing within the <select> element should either be removed or relocated to a valid element that allows text descendants, e.g., an <optgroup> with a <legend> element or <option> elements.
CONSOLE ERROR: An element which is not allowed in the content model of the <optgroup> element was found within an <optgroup> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology. If using disallowed elements for layout structure and styling, consider using the allowed <div> element instead. Any text existing within the <optgroup> element should either be removed or relocated to a valid element that allows text descendants, e.g., the <legend> or <option> elements.
CONSOLE ERROR: An element which is not allowed in the content model of the <optgroup> element was found within an <optgroup> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology. If using disallowed elements for layout structure and styling, consider using the allowed <div> element instead. Any text existing within the <optgroup> element should either be removed or relocated to a valid element that allows text descendants, e.g., the <legend> or <option> elements.
CONSOLE ERROR: An element which is not allowed in the content model of the <optgroup> element was found within an <optgroup> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology. If using disallowed elements for layout structure and styling, consider using the allowed <div> element instead. Any text existing within the <optgroup> element should either be removed or relocated to a valid element that allows text descendants, e.g., the <legend> or <option> elements.
CONSOLE ERROR: An interactive element which is not allowed in the content model of the <legend> element was found within a <legend> element. Interactive elements are not allowed children of a <legend> element when used within an <optgroup> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
CONSOLE ERROR: An interactive element which is not allowed in the content model of the <option> element was found within an <option> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
CONSOLE ERROR: An interactive element which is not allowed in the content model of the <option> element was found within an <option> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
CONSOLE ERROR: An interactive element which is not allowed in the content model of the <option> element was found within an <option> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
CONSOLE ERROR: An interactive element which is not allowed in the content model of the <option> element was found within an <option> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
CONSOLE ERROR: An element which is not allowed in the content model of the <select> element was found within a <select> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology. If using disallowed elements for layout structure and styling, consider using the allowed <div> element instead. Any text existing within the <select> element should either be removed or relocated to a valid element that allows text descendants, e.g., an <optgroup> with a <legend> element or <option> elements.
CONSOLE ERROR: An interactive element which is not allowed in the content model of the <option> element was found within an <option> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
CONSOLE ERROR: An interactive element which is not allowed in the content model of the <option> element was found within an <option> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
CONSOLE ERROR: An interactive element which is not allowed in the content model of the <option> element was found within an <option> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
CONSOLE ERROR: An interactive element which is not allowed in the content model of the <option> element was found within an <option> element. These elements will not consistently be accessible to people navigating by keyboard or using assistive technology.
   
 
 
 
 
 
 
..
 
 
 
..
 
.. ..
 
.. ..
 
.. ..
 
.. ..
 
